+++ l10n_br_hr_holiday/__openerp__.py
@@ -10,7 +10,6 @@
     'website': 'www.kmee.com.br',
     'depends': [
         'hr_holidays',
-        'l10n_br_resource',
         'l10n_br_hr_vacation'
     ],
     'data': [
