+++ l10n_br_hr_holiday/models/__init__.py
@@ -4,4 +4,3 @@
 
 from . import hr_holidays
 from . import hr_holidays_status
-from . import resource_calendar
