+++ l10n_br_hr_holiday/models/hr_holidays.py
@@ -70,3 +70,43 @@ class HrHolidays(models.Model):
     @api.onchange('payroll_discount', 'holiday_status_id')
     def _set_payroll_discount(self):
         self.payroll_discount = self.holiday_status_id.payroll_discount
+
+    @api.multi
+    def get_ocurrences(self, employee_id, data_from, data_to):
+        """Calcular a quantidade de faltas/ocorrencias que devem ser
+        descontadas da folha de pagamento em determinado intervalo de tempo.
+        :param  employee_id: Id do funcionario
+                str data_from: Data inicial do intervalo de tempo.
+                str data_end: Data final do intervalo
+        :return dict {
+                    - int faltas_remuneradas: Quantidade de faltas que devem
+                    ser remuneradas dentro do intervalo passado como parâmetro
+                    - int faltas_nao_remuneradas: Quantidade de faltas que NAO
+                    serao remuneradas. (Descontadas da folha de pagamento)
+        }
+        """
+        faltas = {
+            'faltas_remuneradas': [],
+            'quantidade_dias_faltas_remuneradas': 0,
+            'faltas_nao_remuneradas': [],
+            'quantidade_dias_faltas_nao_remuneradas': 0,
+        }
+        domain = [
+            ('state', '=', 'validate'),
+            ('employee_id', '=', employee_id),
+            ('type', '=', 'remove'),
+            ('date_from', '>=', data_from),
+            ('date_to', '<=', data_to),
+        ]
+        holidays_ids = self.env['hr.holidays'].search(domain)
+
+        for leave in holidays_ids:
+            if leave.payroll_discount:
+                faltas['faltas_nao_remuneradas'].append(leave)
+                faltas['quantidade_dias_faltas_nao_remuneradas'] += \
+                    leave.number_of_days_temp
+            else:
+                faltas['faltas_remuneradas'].append(leave)
+                faltas['quantidade_dias_faltas_remuneradas'] += \
+                    leave.number_of_days_temp
+        return faltas
